student-teacher-meeting-scheduler-api-client/pyproject.toml
[tool.poetry]
name = "student-teacher-meeting-scheduler-api-client"
version = "1.3.0"
description = "A client library for accessing Student-Teacher Meeting Scheduler API"
authors = []
readme = "README.md"
packages = [
    {include = "student_teacher_meeting_scheduler_api_client"},
]
include = ["CHANGELOG.md", "student_teacher_meeting_scheduler_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
